Fully split RegexParser
