+++ logscan/logscan.go
@@ -10,12 +10,9 @@ import (
 	"fmt"
 	"io"
 	"regexp"
-	"slices"
-	"strconv"
 	"strings"
 	"time"
 
-	"github.com/bmatcuk/doublestar/v4"
 	"zgo.at/errors"
 	"zgo.at/follow"
 	"zgo.at/zlog"
@@ -34,61 +31,27 @@ const (
 	excludeRe       = 2
 )
 
-type excludePattern struct {
-	kind    int            // exclude* constant
-	negate  bool           // ! present
-	field   string         // "path", "content_type"
-	pattern string         // ".gif", "*.gif"
-	re      *regexp.Regexp // only if kind=excludeRe
+type LineParser interface {
+	Parse(string) (Line, bool, error)
 }
 
-func processExcludes(exclude []string) ([]excludePattern, error) {
-	// "static" needs to expand to two values.
-	for i, e := range exclude {
-		switch e {
-		case "static":
-			// Note: maybe check if using glob patterns is faster?
-			exclude[i] = `path:re:.*\.(:?js|css|gif|jpe?g|png|svg|ico|web[mp]|mp[34])$`
-			exclude = append(exclude, `content_type:re:^(?:text/(?:css|javascript)|image/(?:png|gif|jpeg|svg\+xml|webp)).*?`)
-		case "html":
-			exclude[i] = "content_type:^text/html.*?"
-		case "redirect":
-			exclude[i] = "status:glob:30[0123]"
-		}
-	}
+type Line interface {
+	Host() string
+	RemoteAddr() string
+	XForwardedFor() string
+	Method() string
+	HTTP() string
+	Path() string
+	Query() string
+	Referrer() string
+	UserAgent() string
+	ContentType() string
+	Status() int
+	Size() int
+	Language() string
 
-	patterns := make([]excludePattern, 0, len(exclude))
-	for _, e := range exclude {
-		var p excludePattern
-		if strings.HasPrefix(e, "!") {
-			p.negate = true
-			e = e[1:]
-		}
-
-		p.field, p.pattern, _ = strings.Cut(e, ":")
-		if !slices.Contains(fields, p.field) {
-			return nil, fmt.Errorf("invalid field %q in exclude pattern %q", p.field, e)
-		}
-		if p.pattern == "" {
-			return nil, fmt.Errorf("no pattern in %q", e)
-		}
-
-		var err error
-		switch {
-		case strings.HasPrefix(p.pattern, "glob:"):
-			p.kind, p.pattern = excludeGlob, p.pattern[5:]
-			_, err = doublestar.Match(p.pattern, "")
-		case strings.HasPrefix(p.pattern, "re:"):
-			p.kind, p.pattern = excludeRe, p.pattern[3:]
-			p.re, err = regexp.Compile(p.pattern)
-		}
-		if err != nil {
-			return nil, fmt.Errorf("invalid exclude pattern: %q: %w", e, err)
-		}
-		patterns = append(patterns, p)
-	}
-
-	return patterns, nil
+	Timing() time.Duration
+	Datetime(scan *Scanner) (time.Time, error)
 }
 
 const (
@@ -182,15 +145,15 @@ func makeNew(format, date, tyme, datetime string, exclude []string) (*Scanner, e
 }
 
 // Line processes a single line.
-func (s *Scanner) Line(ctx context.Context) (Line, error) {
+func (s *Scanner) Line(ctx context.Context) (Line, string, uint64, error) {
 start:
 	var line string
 	select {
 	case <-ctx.Done():
-		return Line{}, io.EOF
+		return nil, "", 0, io.EOF
 	case r := <-s.read:
 		if r.Err != nil {
-			return nil, r.Err
+			return nil, "", 0, r.Err
 		}
 		line = r.String()
 		s.lineno++
@@ -198,7 +161,7 @@ start:
 
 	parsed, excluded, err := s.lp.Parse(line)
 	if err != nil {
-		return nil, err
+		return nil, "", 0, err
 	}
 	// Could use "return Line(ctx) as well, but if many lines are excluded
 	// that will run out of stack space. So just restart the function from
@@ -206,94 +169,6 @@ start:
 	if excluded {
 		goto start
 	}
-	parsed["_line"] = line
-	parsed["_lineno"] = strconv.FormatUint(s.lineno, 10)
 
-	return parsed, nil
-}
-
-func (l Line) exclude(e excludePattern) bool {
-	var m bool
-	switch e.kind {
-	default:
-		m = strings.Contains(l[e.field], e.pattern)
-	case excludeGlob:
-		// We use doublestar instead of filepath.Match() because the latter
-		// doesn't support "**" and "{a,b}" patterns, both of which are very
-		// useful here.
-		m, _ = doublestar.Match(e.pattern, l[e.field])
-	case excludeRe:
-		m = e.re.MatchString(l[e.field])
-	}
-	if e.negate {
-		return !m
-	}
-	return m
-}
-
-func toI(s string) int {
-	n, _ := strconv.Atoi(s) // Regexp only captures \d, so safe to ignore.
-	return n
-}
-func toI64(s string) int64 {
-	n, _ := strconv.ParseInt(s, 10, 64)
-	return n
-}
-func toUi64(s string) uint64 {
-	n, _ := strconv.ParseUint(s, 10, 64)
-	return n
-}
-
-type Line map[string]string
-
-func (l Line) Line() string          { return l["_line"] }
-func (l Line) LineNo() uint64        { return toUi64(l["_lineno"]) }
-func (l Line) Host() string          { return l["host"] }
-func (l Line) RemoteAddr() string    { return l["remote_addr"] }
-func (l Line) XForwardedFor() string { return l["xff"] }
-func (l Line) Method() string        { return l["method"] }
-func (l Line) HTTP() string          { return l["http"] }
-func (l Line) Path() string          { return l["path"] }
-func (l Line) Query() string         { return l["query"] }
-func (l Line) Referrer() string      { return l["referrer"] }
-func (l Line) UserAgent() string     { return l["user_agent"] }
-func (l Line) ContentType() string   { return l["content_type"] }
-func (l Line) Status() int           { return toI(l["status"]) }
-func (l Line) Size() int             { return toI(l["size"]) }
-
-func (l Line) Timing() time.Duration {
-	s, ok := l["timing_sec"]
-	if ok {
-		return time.Duration(toI(s)) * time.Second
-	}
-	s, ok = l["timing_milli"]
-	if ok {
-		return time.Duration(toI64(s)) * time.Millisecond
-	}
-	s, ok = l["timing_micro"]
-	if ok {
-		return time.Duration(toI64(s)) * time.Microsecond
-	}
-	return 0
-}
-
-func (l Line) Datetime(scan *Scanner) (time.Time, error) {
-	date, tyme, datetime := scan.DateFormats()
-
-	s, ok := l["date"]
-	if ok {
-		t, err := time.Parse(date, s)
-		return t.UTC(), err
-	}
-	s, ok = l["time"]
-	if ok {
-		t, err := time.Parse(tyme, s)
-		return t.UTC(), err
-	}
-	s, ok = l["datetime"]
-	if ok {
-		t, err := time.Parse(datetime, s)
-		return t.UTC(), err
-	}
-	return time.Time{}, nil
+	return parsed, line, s.lineno, nil
 }
